refactor: singular
